refactor: updating reactions with new values when addReactions is called
