+++ lib/store/events/actions.dart
@@ -11,8 +11,6 @@ import 'package:syphon/store/events/messages/actions.dart';
 import 'package:syphon/store/events/messages/model.dart';
 import 'package:syphon/store/events/messages/storage.dart';
 import 'package:syphon/store/events/model.dart';
-import 'package:syphon/store/events/reactions/model.dart';
-import 'package:syphon/store/events/redaction/model.dart';
 import 'package:syphon/store/index.dart';
 import 'package:syphon/store/rooms/actions.dart';
 import 'package:syphon/store/rooms/room/model.dart';
@@ -52,36 +50,15 @@ class AddMessages {
   });
 }
 
-class AddReactions {
-  final String? roomId;
-  final List<Reaction>? reactions;
-  AddReactions({this.roomId, this.reactions});
-}
-
 class SetReceipts {
   final String? roomId;
   final Map<String, ReadReceipt>? receipts;
   SetReceipts({this.roomId, this.receipts});
 }
 
-class SaveRedactions {
-  final List<Redaction>? redactions;
-  SaveRedactions({this.redactions});
-}
-
-class LoadReactions {
-  final Map<String, List<Reaction>> reactionsMap;
-  LoadReactions({required this.reactionsMap});
-}
-
 class LoadReceipts {
   final Map<String, Map<String, ReadReceipt>> receiptsMap;
   LoadReceipts({required this.receiptsMap});
-}
-
-class LoadRedactions {
-  final Map<String, Redaction> redactionsMap;
-  LoadRedactions({required this.redactionsMap});
 }
 
 ///
@@ -127,58 +104,6 @@ ThunkAction<AppState> addMessages({
     };
 
 ///
-/// Redact Messages
-///
-/// Redact messages locally throughout all
-/// storage layers
-///
-ThunkAction<AppState> redactEvents({required Room room, List<Redaction> redactions = const []}) {
-  return (Store<AppState> store) async {
-    try {
-      if (redactions.isEmpty) return;
-
-      final messagesCached = store.state.eventStore.messages[room.id] ?? [];
-      final reactionsCached = store.state.eventStore.reactions[room.id] ?? [];
-
-      // create a map of messages for O(1) when replacing O(N)
-      final messagesMap = Map<String, Message>.fromIterable(
-        messagesCached,
-        key: (message) => message.id,
-        value: (message) => message,
-      );
-
-      // create a map of messages for O(1) when replacing O(N)
-      final reactionsMap = Map<String, Reaction>.fromIterable(
-        reactionsCached,
-        key: (message) => message.id,
-        value: (message) => message,
-      );
-
-      final messages = <Message>[];
-      final reactions = <Reaction>[];
-
-      for (final redaction in redactions) {
-        if (messagesMap.containsKey(redaction.redactId)) {
-          messages.add(messagesMap[redaction.redactId]!.copyWith(body: null));
-        }
-        if (reactionsMap.containsKey(redaction.redactId)) {
-          reactions.add(reactionsMap[redaction.redactId]!.copyWith(body: null));
-        }
-      }
-
-      // add messages back to cache having been redacted
-      store.dispatch(addMessages(room: room, messages: messages));
-      store.dispatch(addReactions(reactions: reactions));
-
-      // save redactions to cold storage
-      store.dispatch(SaveRedactions(redactions: redactions));
-    } catch (error) {
-      printError('[deleteMessage] $error');
-    }
-  };
-}
-
-///
 /// Add Messages Decrypted
 ///
 /// Saves in memory only version of the decrypted message
@@ -194,23 +119,6 @@ ThunkAction<AppState> addMessagesDecrypted({
       return store.dispatch(
         AddMessagesDecrypted(roomId: room.id, messages: messages, outbox: outbox),
       );
-    };
-
-ThunkAction<AppState> addReactions({
-  List<Reaction> reactions = const [],
-}) =>
-    (Store<AppState> store) {
-      if (reactions.isEmpty) return;
-      return store.dispatch(AddReactions(reactions: reactions));
-    };
-
-ThunkAction<AppState> setRedactions({
-  String? roomId,
-  List<Redaction>? redactions,
-}) =>
-    (Store<AppState> store) {
-      if (redactions!.isEmpty) return;
-      store.dispatch(SaveRedactions(redactions: redactions));
     };
 
 ThunkAction<AppState> setReceipts({
@@ -525,28 +433,6 @@ ThunkAction<AppState> deleteMessage({required Message message, required Room roo
       final messageDeleted = message.copyWith(body: '');
 
       return store.dispatch(DeleteMessage(room: room, message: messageDeleted));
-    } catch (error) {
-      printError('[deleteMessage] $error');
-    }
-  };
-}
-
-///
-/// Redact Event
-///
-/// Only use when you're sure no temporary events
-/// can be removed first (like failed or pending sends)
-///
-ThunkAction<AppState> redactEvent({Room? room, Event? event}) {
-  return (Store<AppState> store) async {
-    try {
-      await MatrixApi.redactEvent(
-        trxId: DateTime.now().millisecond.toString(),
-        accessToken: store.state.authStore.user.accessToken,
-        homeserver: store.state.authStore.user.homeserver,
-        roomId: room!.id,
-        eventId: event!.id,
-      );
     } catch (error) {
       printError('[deleteMessage] $error');
     }
