+++ lib/store/events/reactions/actions.dart
@@ -3,11 +3,31 @@ import 'package:redux/redux.dart';
 import 'package:redux_thunk/redux_thunk.dart';
 import 'package:syphon/global/libs/matrix/index.dart';
 import 'package:syphon/global/print.dart';
-import 'package:syphon/store/events/actions.dart';
 import 'package:syphon/store/events/messages/model.dart';
+import 'package:syphon/store/events/reactions/model.dart';
+import 'package:syphon/store/events/redaction/actions.dart';
 import 'package:syphon/store/index.dart';
 import 'package:syphon/store/rooms/actions.dart';
 import 'package:syphon/store/rooms/room/model.dart';
+
+class AddReactions {
+  final String? roomId;
+  final List<Reaction>? reactions;
+  AddReactions({this.roomId, this.reactions});
+}
+
+class LoadReactions {
+  final Map<String, List<Reaction>> reactionsMap;
+  LoadReactions({required this.reactionsMap});
+}
+
+ThunkAction<AppState> addReactions({
+  List<Reaction> reactions = const [],
+}) =>
+    (Store<AppState> store) {
+      if (reactions.isEmpty) return;
+      return store.dispatch(AddReactions(reactions: reactions));
+    };
 
 ThunkAction<AppState> toggleReaction({
   Room? room,
@@ -23,7 +43,7 @@ ThunkAction<AppState> toggleReaction({
     if (reaction == null) {
       store.dispatch(sendReaction(message: message, room: room, emoji: emoji));
     } else {
-      store.dispatch(redactEvent(event: reaction, room: room));
+      store.dispatch(sendRedaction(event: reaction, room: room));
     }
   };
 }
